extract db layer into db.go. convert everyting to use existing graph.Properteis api
